BUGFIX: be more precise when building SSM Parameter prefixes for searching values
We were splitting on "." and taking the first two bits, so if you had "access_foo.test.BAR" you would get a prefix of "access_foo.test".  This screws you up when you also have parameters in the parameter space with prefixes like "access_foo.test0.BAR".   Now we add a final "." to the prefix to fix that.