BUGFIX: task list checks the returned tasks for desired task_type.
The bug occured because a resource was once tagged a standalone task
and, although it was removed in recent revisions, `task list` collected
the pk as long as at least one revision was tagged that way. Since
the `StandaloneTaskManager.get` returns the latest task revision (or a
specific revision), we can check if the task_type is available and
if it still matches the desired task_type. If it doesn't match, we skip
that task.