Add plan command to service to show diff between deployfish and AWS.
Deployfish models inherit from core.models.abtract.Model and that
comes with render_for_diff and diff. Inspired, I updating the deployfish
Service model to match the AWS Service model so jsondiff had a more
accurate comparison.

Since a jinja2 environment was not set for the renderer, and not wanting
to affect existing templates, I manually trimmed all whitespace and newlines,
and explicitly add in newlines for my printouts. The template is
basically a recursive macro that transverses the json with changes and
show original data on updates.

A utility macro was added for adding spaces for indents.

Additional notes:
- VpcConfigurationMixin.get_vpc_configuration for ECS adapters was updated
  to sorted lists values so that they wouldn't generate changes, this
  affected subnets and security_groups
- TaskDefinition tags can be set after the deployfish Service object has
  been created. A timestamp of the current time is usually added during
  models.Servce.render, but tags for Services tasks were not included.
  Since service.tasks are loaded in separately from deployfish.yml, I
  updated models.Service.new to recheck if taskDefinition was added
  and then looped through intance.helper_tasks to add the task tags
  with the value as TBD. Updates will always cause the TaskDefinition
  to increment the revision version.
- Timestamp in TaskDefinition will always be seen as a change also.
- You can use the -d to display line markers and nested depth to debug
  the recursive macro.

Not sure how robust the new recursive macro is for the template. But
it's nice start.