Remove commented out requiresCompatibilities and restore roleArn.
Didn't seem necessary to override requiresCompatibilities with
capabilities because AWS supplies it, and requireCapabilities
doesn't always equal capabilities in our Task Description json.

I had removed roleArn because it didn't look like it was useful,
however just because I haven't encountered a need doesn't mean
it's not used for something else that I'm not aware of. Restoring
it in case it is used for something. It doesn't affect the new
feature.