BUGFIX: modified some render_for_diff attributes and handle a list update that is really an insert plan template.
Deleted ScalableTargetARN from ScalableTarget.render_for_diff
since I don't think we have control of that value, and it doesn't
provide much information.

Task definition template was missing requiresCapabilities for an
older service due to taskDefinitionArn not in the render data. I
moved it out of the conditional.

The old service also revealed a unhandled section in service plan
template macro: a list update that is really a list insert in
disguise. It seems that service AWS obj had empty list of tags
but the deplofish object contained a tag. (This might relate our
terraform files because our more recent services have that tag saved.)
Jsondiff registered this difference as an update to the list
and not an insert. So I looped through the update list of dicts and set
the mode to insert to properly display the new list items.