Fixed that annoying syntax warning about an invalid escape sequence.
Python is going to turn this warning into a syntax error soon, which is why it's now a warning instead of just being ignored.

Of the two available solutions (escape the \ with another \, or make the docstring a raw string, I went with raw string so that anyone copying the example code directly out of the .py file (rather than the docs) won't get an invalid SSH ProxyCommand line.