Cleaned up the error message when `deploy tunnel` fails due to a missing deployfish.yml
Because the tunnel command does its config loading outside of the normal workflow for that, it wasn't catching the exception thrown from Config.load_config().

I also added cement and tabulate to the requirements.txt, as they were only in the setup.cfg file, so they wouldn't install themselves in your development virtualenv.